Create README.md
The Gym Workout Planner and Tracker is aimed at developing a Desktop Java
Application that will monitor the workouts of participants under the guidance of
specialized coaches. The primary goal of this application is to help participants achieve
their fitness goals by providing tailored workout programs and nutrition plans. The
application will also act as a platform for coaches to manage and track the progress of
their participants.